feat: fix cmake list error

+++ Team41/Code41/src/spa/src/QPS/Optimizer/ClauseDepGraph.cpp
@@ -17,7 +17,7 @@ void ClauseDepGraph::registerClause(TempClause qc) {
         synonymToClauses[NO_SYNONYM].push_back(qc);
         return;
     }
-    int len = synonyms.size();
+    int len = (int) synonyms.size();
     for (int i = 0; i < len; i++) {
         if (i != 0) graph.addEdge(synonyms[i-1].getLabel(), synonyms[i].getLabel());
         synonymToClauses[synonyms[i].getLabel()].push_back(qc);
@@ -31,7 +31,7 @@ ClauseGroups ClauseDepGraph::split() {
     // populate clause groups based on components
     ClauseGroups clauseGroups(N, pkbAdapter);
     for (int gid = 0; gid < N; gid++) { // loop all groups of synonyms
-        int len = groups[gid].size();
+        int len = (int) groups[gid].size();
         for (int cid = 0; cid < len; cid++) { // loop one group of syns
             for (const TempClause &cl: getClausesOfSyn(groups[gid][cid])) {
                 clauseGroups.addClause(gid, cl);
